requirements.txt
openai_whisper==20240930
pydub==0.25.1
SpeechRecognition==3.10.4
streamlit==1.39.0
textblob==0.17.1
